Update hide_gm_rolls.js
I'm still learning what is and isn't possible with this stuff.

Really what I was trying to solve for it seems is to hide the chat messages coming from a hidden token when there is no defined message destination or roll mode.

I still want specifically chosen roll modes to apply but for anything like the combat turn tracker that I cannot specify, it should hide if the token is hidden.